+++ packages/spacecat-shared-data-access/src/index.d.ts
@@ -422,6 +422,12 @@ export interface Configuration {
   getHandler: (type) => object;
 
   /**
+   * Retrieves the slack roles configuration.
+   * @returns {object} The slack roles configuration.
+   */
+  getSlackRoles: () => object;
+
+  /**
    * Return true if a handler type is enabled for an organization.
    * @param type handler type
    * @param org organization
@@ -884,7 +890,7 @@ interface DataAccessConfig {
   indexNameAllOrganizationsByImsOrgId: string,
   indexNameAllImportJobsByStatus: string,
   indexNameAllImportJobsByDateRange: string,
-  indexNameAllImportUrlsByJobIdAndStatus: string,
+  indexNameImportUrlsByJobIdAndStatus: string,
   indexNameApiKeyByHashedApiKey: string,
   pkAllSites: string;
   pkAllLatestAudits: string;