+++ packages/spacecat-shared-data-access/src/models/site/config.js
@@ -57,14 +57,14 @@ export const Config = (data = {}) => {
   const state = { ...validConfig };
   const self = { state };
   self.getSlackConfig = () => state.slack;
-  self.getSlackMentions = (type) => state?.handlers[type]?.mentions?.slack;
-  self.getHandlerConfig = (type) => state?.handlers[type];
+  self.getSlackMentions = (type) => state?.handlers?.[type]?.mentions?.slack;
+  self.getHandlerConfig = (type) => state?.handlers?.[type];
   self.getHandlers = () => state.handlers;
   self.getImports = () => state.imports;
-  self.getExcludedURLs = (type) => state?.handlers[type]?.excludedURLs;
-  self.getManualOverwrites = (type) => state?.handlers[type]?.manualOverwrites;
-  self.getFixedURLs = (type) => state?.handlers[type]?.fixedURLs;
-  self.getIncludedURLs = (type) => state?.handlers[type]?.includedURLs;
+  self.getExcludedURLs = (type) => state?.handlers?.[type]?.excludedURLs;
+  self.getManualOverwrites = (type) => state?.handlers?.[type]?.manualOverwrites;
+  self.getFixedURLs = (type) => state?.handlers?.[type]?.fixedURLs;
+  self.getIncludedURLs = (type) => state?.handlers?.[type]?.includedURLs;
 
   self.updateSlackConfig = (channel, workspace, invitedUserCount) => {
     state.slack = {