+++ packages/spacecat-shared-data-access/src/service/import-url/accessPatterns.js
@@ -99,11 +99,11 @@ export const getImportUrlsByJobIdAndStatus = async (dynamoClient, config, log, j
 
 /**
  * Get Import Urls by Job ID, if no urls exist an empty array is returned.
- * @param {DynamoClient} dynamoClient
- * @param {Object} config
- * @param {Logger} log
- * @param {string} jobId
- * @returns {Promise<ImportUrlDto[]> | []}
+ * @param {DynamoDbClient} dynamoClient - The DynamoDB client.
+ * @param {DataAccessConfig} config - The data access config.
+ * @param {Logger} log - The logger.
+ * @param {string} jobId - The ID of the import job.
+ * @returns {Promise<ImportUrl[]>}
  */
 export const getImportUrlsByJobId = async (dynamoClient, config, log, jobId) => {
   const items = await dynamoClient.query({
@@ -116,4 +116,38 @@ export const getImportUrlsByJobId = async (dynamoClient, config, log, jobId) => 
   });
 
   return items ? items.map((item) => ImportUrlDto.fromDynamoItem(item)) : [];
+};
+
+/**
+ * Remove all given import URLs.
+ * @param {DynamoDbClient} dynamoClient - The DynamoDB client.
+ * @param {DataAccessConfig} config - The data access config.
+ * @param {ImportUrl[]} urls - The import URLs to remove.
+ * @return {Promise<void>} A promise that resolves when all URLs have been removed.
+ */
+async function removeUrls(dynamoClient, config, urls) {
+  const removeUrlPromises = urls.map((url) => dynamoClient.removeItem(
+    config.tableNameImportUrls,
+    { id: url.getId() },
+  ));
+
+  await Promise.all(removeUrlPromises);
+}
+
+/**
+ * Remove all URLs associated with an import job.
+ * @param {DynamoDbClient} dynamoClient - The DynamoDB client.
+ * @param {DataAccessConfig} config - The data access config.
+ * @param {Logger} log - The logger.
+ * @param {string} jobId - The ID of the import job.
+ * @return {Promise<void>} A promise that resolves when all URLs have been removed.
+ */
+export const removeUrlsForImportJob = async (dynamoClient, config, log, jobId) => {
+  try {
+    const urls = await getImportUrlsByJobId(dynamoClient, config, log, jobId);
+    await removeUrls(dynamoClient, config, urls);
+  } catch (error) {
+    log.error(`Error removing urls for import jobId: ${jobId} : ${error.message}`);
+    throw error;
+  }
 };