+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,31 @@
+# [@adobe/spacecat-shared-google-client-v1.2.13](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.12...@adobe/spacecat-shared-google-client-v1.2.13) (2024-10-19)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#405](https://github.com/adobe/spacecat-shared/issues/405)) ([11072b4](https://github.com/adobe/spacecat-shared/commit/11072b4289db7e18328ed8faff4651fb95a1370d))
+
+# [@adobe/spacecat-shared-google-client-v1.2.12](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.11...@adobe/spacecat-shared-google-client-v1.2.12) (2024-10-12)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#402](https://github.com/adobe/spacecat-shared/issues/402)) ([9a5acba](https://github.com/adobe/spacecat-shared/commit/9a5acba2773b83ce26f4ac97e04d8ad24b00d8ce))
+
+# [@adobe/spacecat-shared-google-client-v1.2.11](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.10...@adobe/spacecat-shared-google-client-v1.2.11) (2024-10-08)
+
+
+### Bug Fixes
+
+* gracefully handle api call in case site url is not defined ([#397](https://github.com/adobe/spacecat-shared/issues/397)) ([45924c8](https://github.com/adobe/spacecat-shared/commit/45924c8ef341b238647cb10ea927ab7c634007d2))
+
+# [@adobe/spacecat-shared-google-client-v1.2.10](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.9...@adobe/spacecat-shared-google-client-v1.2.10) (2024-10-05)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#394](https://github.com/adobe/spacecat-shared/issues/394)) ([ad165f5](https://github.com/adobe/spacecat-shared/commit/ad165f5fcf9c0ab17d2e901e7a5c0ed3c39dcb51))
+
 # [@adobe/spacecat-shared-google-client-v1.2.9](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.8...@adobe/spacecat-shared-google-client-v1.2.9) (2024-09-28)